initial implementation of new download-manager. UNSTABLE

+++ app/src/main/java/org/opengapps/app/card/DownloadCard.java
@@ -241,7 +241,7 @@ public class DownloadCard extends CardView {
      * This method loads the id and completes the necessary steps/shows the progress
      */
     public void restoreDownloadProgress() {
-        Long id = prefs.getLong("running_download_id", 0);
+        int id = prefs.getInt("running_download_id", 0);
         if (id != 0) {
             DownloadProgressView progress = (DownloadProgressView) findViewById(R.id.progress_view);
             progress.show(id, fragment);
